docs/svc_controle_coherence/requirements.txt
Flask==3.0.3
flask-swagger-ui==4.11.1
lxml==5.2.2
MarkupSafe==2.1.5
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
distro
setuptools